requirements.txt
fastapi
fastapi.responses
uvicorn
mysql.connector
re